Issue #1 resolved.
Added UpdateHandler to sprite. It will help to move using new movement vector which can be set after some period of time passing.
Currently I add only updater with time. But it can be extended to avoid obstacles and creating path to your unit. But issue can be closed because all issue notifications will be closed with toward implementation process.